+++ packages/solid/src/index.ts
@@ -1,18 +1,5 @@
-import type {
-  FormState as ModularForm,
-  FieldState as ModularField,
-  FieldValue as ModularValue,
-  FieldValues as ModularValues,
-  FieldArrayState as ModularFieldArray,
-} from './types';
-export type {
-  ModularForm,
-  ModularField,
-  ModularValue,
-  ModularValues,
-  ModularFieldArray,
-};
 export * from './components';
 export * from './methods';
 export * from './primitives';
+export * from './types';
 export * from './validation';