+++ packages/website/src/components/FormFooter.tsx
@@ -1,8 +1,8 @@
-import { ModularForm, reset } from '@modular-forms/solid';
+import { FormState, reset } from '@modular-forms/solid';
 import { ActionButton } from './ActionButton';
 
 type FormFooterProps = {
-  of: ModularForm<any>;
+  of: FormState<any>;
 };
 
 /**